+++ src/api/fetchCategories.spec.ts
@@ -6,11 +6,7 @@ import { fetchCategories, fetchCategory } from ".";
 beforeAll(() => server.listen());
 afterAll(() => server.close());
 
-// FIXME: sdkとnextをアップグレードしたら壊れた
-// serviceDomain or endpoint may be wrong.
-// Details: FetchError: request to https://food-blog.microcms.io/api/v1/categories?limit=100 failed, reason: Invalid character in header field name
-
-describe.skip("utils/fetcher/fetchCategories", () => {
+describe("utils/fetcher/fetchCategories", () => {
   const testCategories = Object.values(mockCategories);
 
   describe("fetchCategories", () => {
