+++ src/api/fetchTags.spec.ts
@@ -6,11 +6,7 @@ import { fetchTag, fetchTags } from ".";
 beforeAll(() => server.listen());
 afterAll(() => server.close());
 
-// FIXME: sdkとnextをアップグレードしたら壊れた
-// serviceDomain or endpoint may be wrong.
-// Details: FetchError: request to https://food-blog.microcms.io/api/v1/tags?limit=100 failed, reason: Invalid character in header field name
-
-describe.skip("utils/fetcher/fetchTags", () => {
+describe("utils/fetcher/fetchTags", () => {
   const testTags = Object.values(mockTags);
 
   describe("fetchTags", () => {
